Move BallerinaMetadataReader out of exported packages
